saveP2PBlockToBlockSync func moved to p2p.go
