+++ block/p2p.go
@@ -80,3 +80,17 @@ func (m *Manager) gossipBlock(ctx context.Context, block types.Block, commit typ
 
 	return nil
 }
+
+// This function adds the block to blocksync store to enable P2P retrievability
+func (m *Manager) saveP2PBlockToBlockSync(block *types.Block, commit *types.Commit) error {
+	gossipedBlock := p2p.BlockData{Block: *block, Commit: *commit}
+	gossipedBlockBytes, err := gossipedBlock.MarshalBinary()
+	if err != nil {
+		return fmt.Errorf("marshal binary: %w: %w", err, ErrNonRecoverable)
+	}
+	err = m.P2PClient.SaveBlock(context.Background(), block.Header.Height, gossipedBlockBytes)
+	if err != nil {
+		m.logger.Error("Adding  block to blocksync store.", "err", err, "height", gossipedBlock.Block.Header.Height)
+	}
+	return nil
+}
